extract features to hdf5

+++ audioexplorer/onsets.py
@@ -9,7 +9,7 @@ class OnsetDetector(object):
                  hop: int = 256,
                  onset_detector_type: str = 'hfc',
                  onset_threshold: float = 0.01,
-                 onset_silence_threshold: int = -90,
+                 onset_silence_threshold: float = -90,
                  min_duration_s: float = 0.15):
         self.hop = hop
         self.onset_detector = onset(onset_detector_type, nfft, hop, fs)
